Move Linux UHID backend test hooks to tests
Extract test-only Linux UHID backend hooks and syscall/XTest fakes out of src/ into a dedicated test translation unit (tests/fixtures/linux_backend_test_hooks.cpp) and move the header into tests/fixtures/include. Remove the LIBVIRTUALHID_ENABLE_TEST_HOOKS compile definition from src/CMakeLists.txt so production builds no longer expose test hooks. Update tests/CMakeLists.txt to collect test sources into a variable, conditionally add the Linux test hooks and X11/XTest links on Linux, and wire up compile defines/includes only when available. Add platform helpers and per-OS test base classes in fixtures (fixtures.cpp / fixtures.hpp). Update CI to install X11/XTest and sdl2-jstest, load uhid/uinput modules and adjust permissions for device nodes. Clarify README wording about Linux integration tests and their failure conditions when device nodes or discovery tools are unavailable. These changes keep test-only machinery out of production code and make the test harness platform-aware.
