+++ tests/fixtures/linux_backend_test_hooks.cpp
@@ -1,0 +1,1013 @@
+/**
+ * @file tests/fixtures/linux_backend_test_hooks.cpp
+ * @brief Linux backend test hook definitions.
+ */
+
+// standard includes
+#include <algorithm>
+#include <atomic>
+#include <cerrno>
+#include <chrono>
+#include <cstddef>
+#include <cstring>
+#include <string>
+#include <thread>
+#include <utility>
+#include <vector>
+
+// platform includes
+#if defined(__linux__)
+  #include <fcntl.h>
+  #ifndef __user
+    #define __user
+  #endif
+  #include <linux/uhid.h>
+  #include <linux/uinput.h>
+  #include <poll.h>
+  #include <sys/ioctl.h>
+  #include <sys/socket.h>
+  #include <unistd.h>
+
+  #if defined(LIBVIRTUALHID_HAVE_XTEST)
+    #include <X11/extensions/XTest.h>
+    #include <X11/Xlib.h>
+  #endif
+#endif
+
+// local includes
+#include "fixtures/linux_backend_test_hooks.hpp"
+
+#if defined(__linux__)
+namespace lvh::detail::test {
+  namespace {
+
+    struct LinuxTestSyscalls {
+      bool override_access = false;
+      int access_result = 0;
+      bool override_open = false;
+      int open_result = 100000;
+      bool override_write = false;
+      std::atomic_int write_call_count = 0;
+      int fail_write_call = -1;
+      int short_write_call = -1;
+      std::size_t short_write_size = 1;
+      bool override_ioctl = false;
+      std::atomic_int ioctl_call_count = 0;
+      int fail_ioctl_call = -1;
+      bool override_poll = false;
+      std::atomic_int poll_call_count = 0;
+      std::vector<int> poll_results;
+      std::vector<short> poll_revents;
+      std::vector<int> poll_errors;
+      bool override_read = false;
+      std::atomic_int read_call_count = 0;
+      std::vector<std::ptrdiff_t> read_results;
+      std::vector<int> read_errors;
+      uhid_event read_event {};
+    };
+
+    LinuxTestSyscalls *active_test_syscalls = nullptr;
+
+    class ScopedLinuxTestSyscalls {
+    public:
+      explicit ScopedLinuxTestSyscalls(LinuxTestSyscalls &syscalls):
+          previous_ {active_test_syscalls} {
+        active_test_syscalls = &syscalls;
+      }
+
+      ScopedLinuxTestSyscalls(const ScopedLinuxTestSyscalls &) = delete;
+      ScopedLinuxTestSyscalls &operator=(const ScopedLinuxTestSyscalls &) = delete;
+      ScopedLinuxTestSyscalls(ScopedLinuxTestSyscalls &&) noexcept = delete;
+      ScopedLinuxTestSyscalls &operator=(ScopedLinuxTestSyscalls &&) noexcept = delete;
+
+      ~ScopedLinuxTestSyscalls() {
+        active_test_syscalls = previous_;
+      }
+
+    private:
+      LinuxTestSyscalls *previous_ = nullptr;
+    };
+
+  }  // namespace
+}  // namespace lvh::detail::test
+
+int lvh_linux_test_access(const char *path, int mode) {
+  if (lvh::detail::test::active_test_syscalls != nullptr && lvh::detail::test::active_test_syscalls->override_access) {
+    if (lvh::detail::test::active_test_syscalls->access_result < 0) {
+      errno = EACCES;
+    }
+    return lvh::detail::test::active_test_syscalls->access_result;
+  }
+  return ::access(path, mode);
+}
+
+int lvh_linux_test_open(const char *path, int flags) {
+  if (lvh::detail::test::active_test_syscalls != nullptr && lvh::detail::test::active_test_syscalls->override_open) {
+    if (lvh::detail::test::active_test_syscalls->open_result < 0) {
+      errno = ENOENT;
+      return lvh::detail::test::active_test_syscalls->open_result;
+    }
+    const auto fd = ::open("/dev/null", O_RDWR);
+    if (fd < 0) {
+      errno = EIO;
+    }
+    return fd;
+  }
+  return ::open(path, flags);
+}
+
+std::ptrdiff_t lvh_linux_test_write(int fd, const void *buffer, std::size_t size) {
+  if (lvh::detail::test::active_test_syscalls != nullptr && lvh::detail::test::active_test_syscalls->override_write) {
+    const auto call_count = ++lvh::detail::test::active_test_syscalls->write_call_count;
+    if (lvh::detail::test::active_test_syscalls->fail_write_call == call_count) {
+      errno = EIO;
+      return -1;
+    }
+    if (lvh::detail::test::active_test_syscalls->short_write_call == call_count) {
+      return static_cast<std::ptrdiff_t>(lvh::detail::test::active_test_syscalls->short_write_size);
+    }
+    return static_cast<std::ptrdiff_t>(size);
+  }
+  return static_cast<std::ptrdiff_t>(::write(fd, buffer, size));
+}
+
+int lvh_linux_test_ioctl(int fd, unsigned long request, unsigned long argument) {
+  if (lvh::detail::test::active_test_syscalls != nullptr && lvh::detail::test::active_test_syscalls->override_ioctl) {
+    const auto call_count = ++lvh::detail::test::active_test_syscalls->ioctl_call_count;
+    if (lvh::detail::test::active_test_syscalls->fail_ioctl_call == call_count) {
+      errno = EINVAL;
+      return -1;
+    }
+    return 0;
+  }
+  return ::ioctl(fd, request, argument);
+}
+
+int lvh_linux_test_poll(pollfd *descriptors, nfds_t descriptor_count, int timeout) {
+  if (lvh::detail::test::active_test_syscalls != nullptr && lvh::detail::test::active_test_syscalls->override_poll) {
+    const auto call_index = static_cast<std::size_t>(lvh::detail::test::active_test_syscalls->poll_call_count++);
+    auto result = 0;
+    if (call_index < lvh::detail::test::active_test_syscalls->poll_results.size()) {
+      result = lvh::detail::test::active_test_syscalls->poll_results[call_index];
+    }
+
+    if (descriptor_count > 0) {
+      descriptors[0].revents = 0;
+      if (result > 0 && call_index < lvh::detail::test::active_test_syscalls->poll_revents.size()) {
+        descriptors[0].revents = lvh::detail::test::active_test_syscalls->poll_revents[call_index];
+      }
+    }
+
+    if (result < 0) {
+      errno = call_index < lvh::detail::test::active_test_syscalls->poll_errors.size() ?
+                lvh::detail::test::active_test_syscalls->poll_errors[call_index] :
+                EIO;
+    }
+    return result;
+  }
+  return ::poll(descriptors, descriptor_count, timeout);
+}
+
+std::ptrdiff_t lvh_linux_test_read(int fd, void *buffer, std::size_t size) {
+  if (lvh::detail::test::active_test_syscalls != nullptr && lvh::detail::test::active_test_syscalls->override_read) {
+    const auto call_index = static_cast<std::size_t>(lvh::detail::test::active_test_syscalls->read_call_count++);
+    auto result = std::ptrdiff_t {0};
+    if (call_index < lvh::detail::test::active_test_syscalls->read_results.size()) {
+      result = lvh::detail::test::active_test_syscalls->read_results[call_index];
+    }
+
+    if (result < 0) {
+      errno = call_index < lvh::detail::test::active_test_syscalls->read_errors.size() ?
+                lvh::detail::test::active_test_syscalls->read_errors[call_index] :
+                EIO;
+      return result;
+    }
+
+    if (result > 0) {
+      const auto bytes = std::min<std::size_t>(static_cast<std::size_t>(result), std::min(size, sizeof(uhid_event)));
+      std::memcpy(buffer, &lvh::detail::test::active_test_syscalls->read_event, bytes);
+      return static_cast<std::ptrdiff_t>(bytes);
+    }
+
+    return result;
+  }
+  return static_cast<std::ptrdiff_t>(::read(fd, buffer, size));
+}
+
+#if defined(LIBVIRTUALHID_HAVE_XTEST)
+Display *lvh_linux_test_x_open_display(const char *) {
+  return reinterpret_cast<Display *>(0x1);
+}
+
+int lvh_linux_test_x_close_display(Display *) {
+  return 0;
+}
+
+Bool lvh_linux_test_xtest_query_extension(Display *, int *, int *, int *, int *) {
+  return True;
+}
+
+KeyCode lvh_linux_test_x_keysym_to_keycode(Display *, KeySym keysym) {
+  return keysym == NoSymbol ? 0 : 1;
+}
+
+int lvh_linux_test_x_flush(Display *) {
+  return 0;
+}
+
+int lvh_linux_test_default_screen(Display *) {
+  return 0;
+}
+
+int lvh_linux_test_display_width(Display *, int) {
+  return 1920;
+}
+
+int lvh_linux_test_display_height(Display *, int) {
+  return 1080;
+}
+
+int lvh_linux_test_xtest_fake_key_event(Display *, unsigned int, Bool, unsigned long) {
+  return 1;
+}
+
+int lvh_linux_test_xtest_fake_button_event(Display *, unsigned int, Bool, unsigned long) {
+  return 1;
+}
+
+int lvh_linux_test_xtest_fake_motion_event(Display *, int, int, int, unsigned long) {
+  return 1;
+}
+
+int lvh_linux_test_xtest_fake_relative_motion_event(Display *, int, int, unsigned long) {
+  return 1;
+}
+#endif
+
+#define access lvh_linux_test_access
+#define ioctl lvh_linux_test_ioctl
+#define open lvh_linux_test_open
+#define poll lvh_linux_test_poll
+#define read lvh_linux_test_read
+#define write lvh_linux_test_write
+
+#if defined(LIBVIRTUALHID_HAVE_XTEST)
+  #define DefaultScreen lvh_linux_test_default_screen
+  #define DisplayHeight lvh_linux_test_display_height
+  #define DisplayWidth lvh_linux_test_display_width
+  #define XCloseDisplay lvh_linux_test_x_close_display
+  #define XFlush lvh_linux_test_x_flush
+  #define XKeysymToKeycode lvh_linux_test_x_keysym_to_keycode
+  #define XOpenDisplay lvh_linux_test_x_open_display
+  #define XTestFakeButtonEvent lvh_linux_test_xtest_fake_button_event
+  #define XTestFakeKeyEvent lvh_linux_test_xtest_fake_key_event
+  #define XTestFakeMotionEvent lvh_linux_test_xtest_fake_motion_event
+  #define XTestFakeRelativeMotionEvent lvh_linux_test_xtest_fake_relative_motion_event
+  #define XTestQueryExtension lvh_linux_test_xtest_query_extension
+#endif
+
+#define create_platform_backend create_platform_backend_for_linux_backend_test_hooks
+#include "../../src/platform/linux/uhid_backend.cpp"
+#undef create_platform_backend
+
+#if defined(LIBVIRTUALHID_HAVE_XTEST)
+  #undef XTestQueryExtension
+  #undef XTestFakeRelativeMotionEvent
+  #undef XTestFakeMotionEvent
+  #undef XTestFakeKeyEvent
+  #undef XTestFakeButtonEvent
+  #undef XOpenDisplay
+  #undef XKeysymToKeycode
+  #undef XFlush
+  #undef XCloseDisplay
+  #undef DisplayWidth
+  #undef DisplayHeight
+  #undef DefaultScreen
+#endif
+
+#undef write
+#undef read
+#undef poll
+#undef open
+#undef ioctl
+#undef access
+
+namespace lvh::detail::test {
+  namespace {
+
+    constexpr auto fake_fd = 100000;
+
+    int open_test_fd() {
+      return ::open("/dev/null", O_RDWR);
+    }
+
+    std::vector<LinuxInputEventRecord> read_input_events_until_eof(int fd) {
+      std::vector<LinuxInputEventRecord> records;
+      input_event event {};
+      while (::read(fd, &event, sizeof(event)) == sizeof(event)) {
+        records.push_back({
+          .type = event.type,
+          .code = event.code,
+          .value = event.value,
+        });
+      }
+      return records;
+    }
+
+    bool write_uhid_event(int fd, const uhid_event &event) {
+      auto *data = reinterpret_cast<const char *>(&event);
+      std::size_t written = 0;
+      while (written < sizeof(event)) {
+        const auto result = ::write(fd, data + written, sizeof(event) - written);
+        if (result <= 0) {
+          return false;
+        }
+        written += static_cast<std::size_t>(result);
+      }
+      return true;
+    }
+
+    bool read_uhid_event(int fd, uhid_event &event) {
+      pollfd descriptor {};
+      descriptor.fd = fd;
+      descriptor.events = POLLIN;
+      const auto poll_result = ::poll(&descriptor, 1, 1000);
+      if (poll_result <= 0 || (descriptor.revents & POLLIN) == 0) {
+        return false;
+      }
+
+      auto *data = reinterpret_cast<char *>(&event);
+      std::size_t read_size = 0;
+      while (read_size < sizeof(event)) {
+        const auto result = ::read(fd, data + read_size, sizeof(event) - read_size);
+        if (result <= 0) {
+          return false;
+        }
+        read_size += static_cast<std::size_t>(result);
+      }
+      return true;
+    }
+
+    void enable_fake_device_syscalls(LinuxTestSyscalls &syscalls) {
+      syscalls.override_access = true;
+      syscalls.override_open = true;
+      syscalls.override_write = true;
+      syscalls.override_ioctl = true;
+    }
+
+    bool wait_for_poll_calls(const LinuxTestSyscalls &syscalls, int expected_calls) {
+      using namespace std::chrono_literals;
+
+      for (auto attempt = 0; attempt < 100; ++attempt) {
+        if (syscalls.poll_call_count.load() >= expected_calls) {
+          return true;
+        }
+        std::this_thread::sleep_for(1ms);
+      }
+      return false;
+    }
+
+    Status run_fake_uhid_read_loop(LinuxTestSyscalls &syscalls, int expected_poll_calls) {
+      syscalls.override_write = true;
+
+      ScopedLinuxTestSyscalls scoped_syscalls {syscalls};
+
+      CreateGamepadOptions options;
+      options.profile = profiles::generic_gamepad();
+
+      const auto fd = open_test_fd();
+      if (fd < 0) {
+        return system_error_status(ErrorCode::backend_failure, "failed to open test file descriptor", errno);
+      }
+
+      UhidGamepad gamepad {fd};
+      if (const auto status = gamepad.create(1, options); !status.ok()) {
+        return status;
+      }
+
+      const auto saw_expected_polls = wait_for_poll_calls(syscalls, expected_poll_calls);
+      const auto close_status = gamepad.close();
+      if (!saw_expected_polls) {
+        return Status::failure(ErrorCode::backend_failure, "fake UHID read loop did not consume the scripted poll calls");
+      }
+      return close_status;
+    }
+
+  }  // namespace
+
+  std::string linux_copy_string_char_buffer(const std::string &source) {
+    char destination[5] {};
+    copy_string(destination, source);
+    return destination;
+  }
+
+  int linux_key_code(KeyboardKeyCode key_code) {
+    return key_code_to_linux(key_code);
+  }
+
+  int linux_mouse_button(MouseButton button) {
+    return mouse_button_to_linux(button);
+  }
+
+  std::uint16_t linux_uhid_bus(BusType bus_type) {
+    return to_uhid_bus(bus_type);
+  }
+
+  std::uint16_t linux_uinput_bus(BusType bus_type) {
+    return to_uinput_bus(bus_type);
+  }
+
+  int linux_absolute_axis(std::int32_t value, std::int32_t limit) {
+    return scale_absolute_axis(value, limit);
+  }
+
+  std::vector<std::uint32_t> linux_decode_utf8(const std::string &text) {
+    return decode_utf8(text);
+  }
+
+  std::string linux_uppercase_hex(std::uint32_t codepoint) {
+    return uppercase_hex(codepoint);
+  }
+
+  KeyboardKeyCode linux_hex_digit_key_code(char digit) {
+    return hex_digit_key_code(digit);
+  }
+
+  int linux_legacy_scroll_steps(std::int32_t distance) {
+    return legacy_scroll_steps(distance);
+  }
+
+  std::size_t linux_uhid_descriptor_limit() {
+    uhid_event event {};
+    return sizeof(event.u.create2.rd_data);
+  }
+
+  std::size_t linux_uhid_input_limit() {
+    uhid_event event {};
+    return sizeof(event.u.input2.data);
+  }
+
+  Status linux_uhid_create_with_descriptor_size(std::size_t descriptor_size) {
+    auto profile = profiles::generic_gamepad();
+    profile.report_descriptor.assign(descriptor_size, 0);
+
+    CreateGamepadOptions options;
+    options.profile = std::move(profile);
+
+    UhidGamepad gamepad {-1};
+    return gamepad.create(1, options);
+  }
+
+  Status linux_uhid_submit_report_size(std::size_t report_size) {
+    UhidGamepad gamepad {-1};
+    return gamepad.submit(std::vector<std::uint8_t>(report_size, 0));
+  }
+
+  Status linux_uhid_submit_after_close() {
+    UhidGamepad gamepad {-1};
+    static_cast<void>(gamepad.close());
+    return gamepad.submit({0});
+  }
+
+  Status linux_uinput_keyboard_create_invalid_fd() {
+    CreateKeyboardOptions options;
+    options.profile = profiles::keyboard();
+
+    UinputKeyboard keyboard {-1};
+    return keyboard.create(1, options);
+  }
+
+  Status linux_uinput_keyboard_submit_invalid_fd(const KeyboardEvent &event) {
+    UinputKeyboard keyboard {-1};
+    return keyboard.submit(event);
+  }
+
+  Status linux_uinput_keyboard_type_text_invalid_fd(const std::string &text) {
+    UinputKeyboard keyboard {-1};
+    return keyboard.type_text({.text = text});
+  }
+
+  Status linux_uinput_keyboard_submit_after_close() {
+    UinputKeyboard keyboard {-1};
+    static_cast<void>(keyboard.close());
+    return keyboard.submit({.key_code = 0x41, .pressed = true});
+  }
+
+  LinuxInputSubmissionResult linux_uinput_keyboard_submit_pipe(const KeyboardEvent &event) {
+    int descriptors[2] {-1, -1};
+    if (::pipe(descriptors) != 0) {
+      return {system_error_status(ErrorCode::backend_failure, "failed to create pipe", errno), {}};
+    }
+
+    UinputKeyboard keyboard {descriptors[1]};
+    auto status = keyboard.submit(event);
+    static_cast<void>(keyboard.close());
+    auto records = read_input_events_until_eof(descriptors[0]);
+    static_cast<void>(::close(descriptors[0]));
+    return {std::move(status), std::move(records)};
+  }
+
+  Status linux_uinput_user_device_invalid_fd() {
+    return write_uinput_user_device(-1, profiles::mouse(), 1);
+  }
+
+  Status linux_uinput_user_device_pipe() {
+    int descriptors[2] {-1, -1};
+    if (::pipe(descriptors) != 0) {
+      return system_error_status(ErrorCode::backend_failure, "failed to create pipe", errno);
+    }
+
+    auto status = write_uinput_user_device(descriptors[1], profiles::mouse(), 1);
+    static_cast<void>(::close(descriptors[0]));
+    static_cast<void>(::close(descriptors[1]));
+    return status;
+  }
+
+  Status linux_uinput_mouse_create_invalid_fd() {
+    CreateMouseOptions options;
+    options.profile = profiles::mouse();
+
+    UinputMouse mouse {-1};
+    return mouse.create(1, options);
+  }
+
+  Status linux_uinput_mouse_submit_invalid_fd(const MouseEvent &event) {
+    UinputMouse mouse {-1};
+    return mouse.submit(event);
+  }
+
+  Status linux_uinput_mouse_submit_after_close() {
+    UinputMouse mouse {-1};
+    static_cast<void>(mouse.close());
+    return mouse.submit({.kind = MouseEventKind::relative_motion, .x = 1, .y = 1});
+  }
+
+  LinuxInputSubmissionResult linux_uinput_mouse_submit_pipe(const MouseEvent &event) {
+    int descriptors[2] {-1, -1};
+    if (::pipe(descriptors) != 0) {
+      return {system_error_status(ErrorCode::backend_failure, "failed to create pipe", errno), {}};
+    }
+
+    UinputMouse mouse {descriptors[1]};
+    auto status = mouse.submit(event);
+    static_cast<void>(mouse.close());
+    auto records = read_input_events_until_eof(descriptors[0]);
+    static_cast<void>(::close(descriptors[0]));
+    return {std::move(status), std::move(records)};
+  }
+
+  LinuxUhidRoundTripResult linux_uhid_socketpair_roundtrip() {
+    LinuxUhidRoundTripResult result;
+    int descriptors[2] {-1, -1};
+    if (::socketpair(AF_UNIX, SOCK_STREAM, 0, descriptors) != 0) {
+      result.create_status = system_error_status(ErrorCode::backend_failure, "failed to create socketpair", errno);
+      result.submit_status = result.create_status;
+      result.close_status = result.create_status;
+      return result;
+    }
+
+    auto profile = profiles::xbox_360();
+    CreateGamepadOptions options;
+    options.profile = profile;
+    options.metadata.stable_id = "linux-uhid-roundtrip";
+
+    UhidGamepad gamepad {descriptors[0]};
+    result.create_status = gamepad.create(7, options);
+
+    uhid_event event {};
+    if (read_uhid_event(descriptors[1], event)) {
+      result.saw_create =
+        event.type == UHID_CREATE2 && event.u.create2.vendor == profile.vendor_id && event.u.create2.product == profile.product_id;
+    }
+
+    gamepad.set_output_callback([&result](const GamepadOutput &output) {
+      ++result.output_callback_count;
+      result.last_output = output;
+    });
+
+    event = {};
+    event.type = UHID_OUTPUT;
+    event.u.output.size = static_cast<__u16>(profile.output_report_size);
+    event.u.output.data[0] = profile.report_id;
+    event.u.output.data[1] = 0x34;
+    event.u.output.data[2] = 0x12;
+    event.u.output.data[3] = 0x78;
+    event.u.output.data[4] = 0x56;
+    static_cast<void>(write_uhid_event(descriptors[1], event));
+
+    event = {};
+    event.type = UHID_GET_REPORT;
+    event.u.get_report.id = 9;
+    static_cast<void>(write_uhid_event(descriptors[1], event));
+    if (read_uhid_event(descriptors[1], event)) {
+      result.saw_get_report_reply = event.type == UHID_GET_REPORT_REPLY && event.u.get_report_reply.id == 9;
+    }
+
+    event = {};
+    event.type = UHID_SET_REPORT;
+    event.u.set_report.id = 10;
+    event.u.set_report.size = static_cast<__u16>(profile.output_report_size);
+    event.u.set_report.data[0] = profile.report_id;
+    event.u.set_report.data[1] = 0x78;
+    event.u.set_report.data[2] = 0x56;
+    event.u.set_report.data[3] = 0x34;
+    event.u.set_report.data[4] = 0x12;
+    static_cast<void>(write_uhid_event(descriptors[1], event));
+    if (read_uhid_event(descriptors[1], event)) {
+      result.saw_set_report_reply = event.type == UHID_SET_REPORT_REPLY && event.u.set_report_reply.id == 10;
+    }
+
+    event = {};
+    event.type = UHID_OPEN;
+    static_cast<void>(write_uhid_event(descriptors[1], event));
+
+    lvh::GamepadState state;
+    state.buttons.set(GamepadButton::a);
+    const auto report = reports::pack_input_report(profile, state);
+    result.submit_status = gamepad.submit(report);
+    if (read_uhid_event(descriptors[1], event)) {
+      result.saw_input = event.type == UHID_INPUT2 && event.u.input2.size == report.size();
+    }
+
+    result.close_status = gamepad.close();
+    if (read_uhid_event(descriptors[1], event)) {
+      result.saw_destroy = event.type == UHID_DESTROY;
+    }
+
+    static_cast<void>(::close(descriptors[1]));
+    return result;
+  }
+
+  LinuxBackendFakeCreationResult linux_backend_create_all_fake_success() {
+    LinuxTestSyscalls syscalls;
+    enable_fake_device_syscalls(syscalls);
+    ScopedLinuxTestSyscalls scoped_syscalls {syscalls};
+
+    LinuxBackendFakeCreationResult result;
+    LinuxUhidBackend backend;
+    result.capabilities = backend.capabilities();
+
+    CreateGamepadOptions gamepad_options;
+    gamepad_options.profile = profiles::xbox_360();
+    gamepad_options.metadata.stable_id = "fake-linux-gamepad";
+    auto gamepad = backend.create_gamepad(1, gamepad_options);
+    result.gamepad_status = gamepad.status;
+    if (gamepad) {
+      result.gamepad_close_status = gamepad.gamepad->close();
+    }
+
+    CreateKeyboardOptions keyboard_options;
+    keyboard_options.profile = profiles::keyboard();
+    keyboard_options.stable_id = "fake-linux-keyboard";
+    auto keyboard = backend.create_keyboard(2, keyboard_options);
+    result.keyboard_status = keyboard.status;
+    if (keyboard) {
+      result.keyboard_close_status = keyboard.keyboard->close();
+    }
+
+    CreateMouseOptions mouse_options;
+    mouse_options.profile = profiles::mouse();
+    mouse_options.stable_id = "fake-linux-mouse";
+    auto mouse = backend.create_mouse(3, mouse_options);
+    result.mouse_status = mouse.status;
+    if (mouse) {
+      result.mouse_close_status = mouse.mouse->close();
+    }
+
+    return result;
+  }
+
+  BackendCapabilities linux_backend_fake_unavailable_capabilities() {
+    LinuxTestSyscalls syscalls;
+    syscalls.override_access = true;
+    syscalls.access_result = -1;
+    ScopedLinuxTestSyscalls scoped_syscalls {syscalls};
+
+    LinuxUhidBackend backend;
+    return backend.capabilities();
+  }
+
+  Status linux_backend_gamepad_fake_open_failure() {
+    LinuxTestSyscalls syscalls;
+    syscalls.override_access = true;
+    syscalls.override_open = true;
+    syscalls.open_result = -1;
+    ScopedLinuxTestSyscalls scoped_syscalls {syscalls};
+
+    LinuxUhidBackend backend;
+
+    CreateGamepadOptions options;
+    options.profile = profiles::xbox_360();
+    return backend.create_gamepad(1, options).status;
+  }
+
+  Status linux_backend_gamepad_fake_create_failure() {
+    LinuxTestSyscalls syscalls;
+    enable_fake_device_syscalls(syscalls);
+    syscalls.fail_write_call = 1;
+    ScopedLinuxTestSyscalls scoped_syscalls {syscalls};
+
+    LinuxUhidBackend backend;
+
+    CreateGamepadOptions options;
+    options.profile = profiles::xbox_360();
+    return backend.create_gamepad(1, options).status;
+  }
+
+  Status linux_backend_keyboard_fake_open_failure() {
+    LinuxTestSyscalls syscalls;
+    syscalls.override_access = true;
+    syscalls.override_open = true;
+    syscalls.open_result = -1;
+    ScopedLinuxTestSyscalls scoped_syscalls {syscalls};
+
+    LinuxUhidBackend backend;
+
+    CreateKeyboardOptions options;
+    options.profile = profiles::keyboard();
+    return backend.create_keyboard(1, options).status;
+  }
+
+  Status linux_backend_keyboard_fake_create_failure() {
+    LinuxTestSyscalls syscalls;
+    enable_fake_device_syscalls(syscalls);
+    syscalls.fail_ioctl_call = 1;
+    ScopedLinuxTestSyscalls scoped_syscalls {syscalls};
+
+    LinuxUhidBackend backend;
+
+    CreateKeyboardOptions options;
+    options.profile = profiles::keyboard();
+    return backend.create_keyboard(1, options).status;
+  }
+
+  Status linux_backend_keyboard_fake_fallback_success() {
+#if defined(LIBVIRTUALHID_HAVE_XTEST)
+    LinuxTestSyscalls syscalls;
+    enable_fake_device_syscalls(syscalls);
+    syscalls.fail_ioctl_call = 1;
+    ScopedLinuxTestSyscalls scoped_syscalls {syscalls};
+
+    LinuxUhidBackend backend;
+
+    CreateKeyboardOptions options;
+    options.profile = profiles::keyboard();
+    auto keyboard = backend.create_keyboard(1, options);
+    if (!keyboard) {
+      return keyboard.status;
+    }
+    return keyboard.keyboard->close();
+#else
+    return Status::failure(ErrorCode::backend_unavailable, "XTest fallback is not enabled");
+#endif
+  }
+
+  Status linux_backend_mouse_fake_open_failure() {
+    LinuxTestSyscalls syscalls;
+    syscalls.override_access = true;
+    syscalls.override_open = true;
+    syscalls.open_result = -1;
+    ScopedLinuxTestSyscalls scoped_syscalls {syscalls};
+
+    LinuxUhidBackend backend;
+
+    CreateMouseOptions options;
+    options.profile = profiles::mouse();
+    return backend.create_mouse(1, options).status;
+  }
+
+  Status linux_backend_mouse_fake_create_failure() {
+    LinuxTestSyscalls syscalls;
+    enable_fake_device_syscalls(syscalls);
+    syscalls.fail_ioctl_call = 1;
+    ScopedLinuxTestSyscalls scoped_syscalls {syscalls};
+
+    LinuxUhidBackend backend;
+
+    CreateMouseOptions options;
+    options.profile = profiles::mouse();
+    return backend.create_mouse(1, options).status;
+  }
+
+  Status linux_backend_mouse_fake_fallback_success() {
+#if defined(LIBVIRTUALHID_HAVE_XTEST)
+    LinuxTestSyscalls syscalls;
+    enable_fake_device_syscalls(syscalls);
+    syscalls.fail_ioctl_call = 1;
+    ScopedLinuxTestSyscalls scoped_syscalls {syscalls};
+
+    LinuxUhidBackend backend;
+
+    CreateMouseOptions options;
+    options.profile = profiles::mouse();
+    auto mouse = backend.create_mouse(1, options);
+    if (!mouse) {
+      return mouse.status;
+    }
+    return mouse.mouse->close();
+#else
+    return Status::failure(ErrorCode::backend_unavailable, "XTest fallback is not enabled");
+#endif
+  }
+
+  Status linux_uhid_submit_fake_write_failure() {
+    LinuxTestSyscalls syscalls;
+    syscalls.override_write = true;
+    syscalls.fail_write_call = 1;
+    ScopedLinuxTestSyscalls scoped_syscalls {syscalls};
+
+    UhidGamepad gamepad {fake_fd};
+    return gamepad.submit({0});
+  }
+
+  Status linux_uhid_submit_fake_short_write() {
+    LinuxTestSyscalls syscalls;
+    syscalls.override_write = true;
+    syscalls.short_write_call = 1;
+    ScopedLinuxTestSyscalls scoped_syscalls {syscalls};
+
+    UhidGamepad gamepad {fake_fd};
+    return gamepad.submit({0});
+  }
+
+  Status linux_uhid_close_fake_write_failure() {
+    LinuxTestSyscalls syscalls;
+    syscalls.override_write = true;
+    syscalls.fail_write_call = 1;
+    ScopedLinuxTestSyscalls scoped_syscalls {syscalls};
+
+    UhidGamepad gamepad {fake_fd};
+    return gamepad.close();
+  }
+
+  Status linux_uhid_close_fake_close_failure() {
+    LinuxTestSyscalls syscalls;
+    syscalls.override_write = true;
+    ScopedLinuxTestSyscalls scoped_syscalls {syscalls};
+
+    UhidGamepad gamepad {fake_fd};
+    return gamepad.close();
+  }
+
+  Status linux_uhid_read_loop_fake_retry_branches() {
+    LinuxTestSyscalls syscalls;
+    syscalls.override_poll = true;
+    syscalls.poll_results = {-1, 0, 1, 1, 1};
+    syscalls.poll_revents = {0, 0, 0, POLLIN, POLLIN};
+    syscalls.poll_errors = {EINTR};
+    syscalls.override_read = true;
+    syscalls.read_results = {-1, 0};
+    syscalls.read_errors = {EAGAIN};
+    return run_fake_uhid_read_loop(syscalls, 5);
+  }
+
+  Status linux_uhid_read_loop_fake_poll_errors() {
+    LinuxTestSyscalls syscall_failure;
+    syscall_failure.override_poll = true;
+    syscall_failure.poll_results = {-1};
+    syscall_failure.poll_errors = {EIO};
+    if (const auto status = run_fake_uhid_read_loop(syscall_failure, 1); !status.ok()) {
+      return status;
+    }
+
+    LinuxTestSyscalls event_failure;
+    event_failure.override_poll = true;
+    event_failure.poll_results = {1};
+    event_failure.poll_revents = {POLLERR};
+    return run_fake_uhid_read_loop(event_failure, 1);
+  }
+
+  Status linux_uhid_read_loop_fake_read_error() {
+    LinuxTestSyscalls syscalls;
+    syscalls.override_poll = true;
+    syscalls.poll_results = {1};
+    syscalls.poll_revents = {POLLIN};
+    syscalls.override_read = true;
+    syscalls.read_results = {-1};
+    syscalls.read_errors = {EIO};
+    return run_fake_uhid_read_loop(syscalls, 1);
+  }
+
+  Status linux_uhid_read_loop_fake_output_without_callback() {
+    LinuxTestSyscalls syscalls;
+    syscalls.override_poll = true;
+    syscalls.poll_results = {1, 1};
+    syscalls.poll_revents = {POLLIN, POLLIN};
+    syscalls.override_read = true;
+    syscalls.read_results = {static_cast<std::ptrdiff_t>(sizeof(uhid_event)), 0};
+    syscalls.read_event.type = UHID_OUTPUT;
+    return run_fake_uhid_read_loop(syscalls, 2);
+  }
+
+  Status linux_uinput_keyboard_create_fake_ioctl_failure(int fail_ioctl_call) {
+    LinuxTestSyscalls syscalls;
+    syscalls.override_write = true;
+    syscalls.override_ioctl = true;
+    syscalls.fail_ioctl_call = fail_ioctl_call;
+    ScopedLinuxTestSyscalls scoped_syscalls {syscalls};
+
+    CreateKeyboardOptions options;
+    options.profile = profiles::keyboard();
+
+    UinputKeyboard keyboard {fake_fd};
+    return keyboard.create(1, options);
+  }
+
+  Status linux_uinput_user_device_fake_short_write() {
+    LinuxTestSyscalls syscalls;
+    syscalls.override_write = true;
+    syscalls.short_write_call = 1;
+    ScopedLinuxTestSyscalls scoped_syscalls {syscalls};
+
+    return write_uinput_user_device(fake_fd, profiles::mouse(), 1);
+  }
+
+  Status linux_uinput_user_device_fake_create_failure() {
+    LinuxTestSyscalls syscalls;
+    syscalls.override_write = true;
+    syscalls.override_ioctl = true;
+    syscalls.fail_ioctl_call = 1;
+    ScopedLinuxTestSyscalls scoped_syscalls {syscalls};
+
+    return write_uinput_user_device(fake_fd, profiles::mouse(), 1);
+  }
+
+  Status linux_uinput_keyboard_submit_fake_write_failure() {
+    LinuxTestSyscalls syscalls;
+    syscalls.override_write = true;
+    syscalls.fail_write_call = 1;
+    syscalls.override_ioctl = true;
+    ScopedLinuxTestSyscalls scoped_syscalls {syscalls};
+
+    UinputKeyboard keyboard {fake_fd};
+    return keyboard.submit({.key_code = 0x41, .pressed = true});
+  }
+
+  Status linux_uinput_keyboard_submit_fake_short_write() {
+    LinuxTestSyscalls syscalls;
+    syscalls.override_write = true;
+    syscalls.short_write_call = 1;
+    syscalls.override_ioctl = true;
+    ScopedLinuxTestSyscalls scoped_syscalls {syscalls};
+
+    UinputKeyboard keyboard {fake_fd};
+    return keyboard.submit({.key_code = 0x41, .pressed = true});
+  }
+
+  Status linux_uinput_keyboard_type_text_fake_success() {
+    LinuxTestSyscalls syscalls;
+    syscalls.override_write = true;
+    syscalls.override_ioctl = true;
+    ScopedLinuxTestSyscalls scoped_syscalls {syscalls};
+
+    UinputKeyboard keyboard {fake_fd};
+    return keyboard.type_text({.text = "A"});
+  }
+
+  Status linux_uinput_keyboard_close_fake_close_failure() {
+    LinuxTestSyscalls syscalls;
+    syscalls.override_ioctl = true;
+    ScopedLinuxTestSyscalls scoped_syscalls {syscalls};
+
+    UinputKeyboard keyboard {fake_fd};
+    return keyboard.close();
+  }
+
+  Status linux_uinput_mouse_create_fake_ioctl_failure(int fail_ioctl_call) {
+    LinuxTestSyscalls syscalls;
+    syscalls.override_write = true;
+    syscalls.override_ioctl = true;
+    syscalls.fail_ioctl_call = fail_ioctl_call;
+    ScopedLinuxTestSyscalls scoped_syscalls {syscalls};
+
+    CreateMouseOptions options;
+    options.profile = profiles::mouse();
+
+    UinputMouse mouse {fake_fd};
+    return mouse.create(1, options);
+  }
+
+  Status linux_uinput_mouse_submit_fake_write_failure(const MouseEvent &event) {
+    LinuxTestSyscalls syscalls;
+    syscalls.override_write = true;
+    syscalls.fail_write_call = 1;
+    syscalls.override_ioctl = true;
+    ScopedLinuxTestSyscalls scoped_syscalls {syscalls};
+
+    UinputMouse mouse {fake_fd};
+    return mouse.submit(event);
+  }
+
+  Status linux_uinput_mouse_submit_fake_short_write(const MouseEvent &event) {
+    LinuxTestSyscalls syscalls;
+    syscalls.override_write = true;
+    syscalls.short_write_call = 1;
+    syscalls.override_ioctl = true;
+    ScopedLinuxTestSyscalls scoped_syscalls {syscalls};
+
+    UinputMouse mouse {fake_fd};
+    return mouse.submit(event);
+  }
+
+}  // namespace lvh::detail::test
+#endif
